add --inputs option to allow specifying input string (we already had --input option to specify file)

+++ csbench_utils.c
@@ -676,6 +676,10 @@ int tmpfile_fd(void) {
         csperror("mkstemp");
         return -1;
     }
+    if (fcntl(fd, F_SETFD, FD_CLOEXEC) == -1) {
+        csperror("fcntl");
+        return false;
+    }
     unlink(path);
     return fd;
 }
